requirements.txt
beautifulsoup4==4.12.2
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.1.0
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
fake-useragent==1.1.3
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
jsmin==3.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pdfkit==1.0.0
pygame==2.3.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
python-apt==2.4.0+ubuntu1
python-dotenv==1.0.0
PyYAML==5.4.1
regex==2023.3.23
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.4.1
systemd-python==234
termcolor==2.3.0
tiktoken==0.3.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
wadllib==1.3.6
zipp==1.0.0
